+++ bindings/nostr-sdk-ffi/bindings-python/examples/client-with-opts.py
@@ -17,10 +17,8 @@ async def main():
             .proxy(proxy))
     client = Client.with_opts(None, opts)
 
-    await client.add_relays([
-        "wss://relay.damus.io",
-        "ws://oxtrdevav64z64yb7x6rjg4ntzqjhedm5b5zjqulugknhzr46ny2qbad.onion"
-    ])
+    await client.add_relay("wss://relay.damus.io")
+    await client.add_relay("ws://oxtrdevav64z64yb7x6rjg4ntzqjhedm5b5zjqulugknhzr46ny2qbad.onion")
     await client.connect()
 
     event = EventBuilder.text_note("Hello from Rust Nostr Python bindings!", []).to_event(keys)