+++ bindings/nostr-sdk-ffi/bindings-python/examples/tor.py
@@ -16,11 +16,9 @@ async def main():
     opts = Options().connection(connection).connection_timeout(timedelta(seconds=60))
     client = Client.with_opts(signer, opts)
 
-    await client.add_relays([
-        "wss://relay.damus.io",
-        "ws://oxtrdevav64z64yb7x6rjg4ntzqjhedm5b5zjqulugknhzr46ny2qbad.onion",
-        "ws://2jsnlhfnelig5acq6iacydmzdbdmg7xwunm4xl6qwbvzacw4lwrjmlyd.onion",
-    ])
+    await client.add_relay("wss://relay.damus.io")
+    await client.add_relay("ws://oxtrdevav64z64yb7x6rjg4ntzqjhedm5b5zjqulugknhzr46ny2qbad.onion")
+    await client.add_relay("ws://2jsnlhfnelig5acq6iacydmzdbdmg7xwunm4xl6qwbvzacw4lwrjmlyd.onion")
     await client.connect()
 
     event = EventBuilder.text_note("Hello from rust-nostr Python bindings!", [])